Alterado para Dar Entrada no veiculo após o cadastro executaEntrada.go

+++ imprimirComprovante/imprimeEntrada.go
@@ -16,7 +16,7 @@ var (
 	modelo string
 )
 
-func ImprimirEntrada(pla, modelo, cor string, id int) {
+func ImprimirEntrada(pla, modelo, cor string, id int64) {
 
 	file, erro := os.Create("pla.txt")
 	if erro != nil {
